requirements.txt
six
redis
tornado
tabulate
pycurl
pyyaml
msgpack-python
